+++ R/select_columns.R
@@ -117,9 +117,13 @@ to_sql.relop_select_columns <- function (x,
   }
   using <- calc_using_relop_select_columns(x,
                                            using = using)
+  qlimit = limit
+  if(!getDBOption(db, "use_pass_limit", TRUE)) {
+    qlimit = NULL
+  }
   subsql_list <- to_sql(x$source[[1]],
                         db = db,
-                        limit = limit,
+                        limit = qlimit,
                         source_limit = source_limit,
                         indent_level = indent_level + 1,
                         tnum = tnum,